refactor(tui): split format.go below 1000 lines
Move grid and related formatters into sibling files in package tui.
No behavior change.

Related: #652
