+++ internal/tui/format.go
@@ -2,16 +2,12 @@ package tui
 
 import (
 	"fmt"
-	"image/color"
 	"sort"
 	"strings"
 	"time"
-	"unicode/utf8"
 
 	lipgloss "charm.land/lipgloss/v2"
-	"charm.land/lipgloss/v2/table"
 
-	"github.com/douglasdemoura/chroncal/internal/event"
 	"github.com/douglasdemoura/chroncal/internal/tui/oklch"
 )
 
@@ -20,343 +16,6 @@ const (
 	totalHours          = 24
 	timeLabelWidth      = 8
 )
-
-// formatTimeColumn returns a fixed-width label for an event's time slot.
-// The width matches "15:04-15:04" (11 chars) so titles line up across
-// all-day events, events with only a start time, and events with a range.
-func formatTimeColumn(ev event.Event) string {
-	switch {
-	case ev.AllDay:
-		return "           "
-	case ev.EndTime.IsZero():
-		return ev.StartTime.Local().Format("15:04") + "      "
-	default:
-		return ev.StartTime.Local().Format("15:04") + "-" + ev.EndTime.Local().Format("15:04")
-	}
-}
-
-// formatTimeColumnMulti returns the time-column label for an event on a
-// specific day within its span. Start day shows "HH:MM→     ", last day
-// shows "     →HH:MM", middle days are blank. All-day and single-day
-// events fall back to formatTimeColumn.
-func formatTimeColumnMulti(ev event.Event, dayIndex, totalDays int) string {
-	if ev.AllDay || totalDays <= 1 {
-		return formatTimeColumn(ev)
-	}
-	switch dayIndex {
-	case 1:
-		return ev.StartTime.Local().Format("15:04") + "→     "
-	case totalDays:
-		return "     →" + ev.EndTime.Local().Format("15:04")
-	default:
-		return "           "
-	}
-}
-
-// effectiveStartOnDay returns the event's start time as experienced on a
-// given day. That is the original start for its first day. It is midnight
-// of that day for any continuation day. Used to sort multi-day events
-// (which stay live from 00:00 on continuation days) above single-day events.
-func effectiveStartOnDay(ev event.Event, day time.Time, dayIndex int) time.Time {
-	if dayIndex == 1 {
-		return ev.StartTime
-	}
-	return time.Date(day.Year(), day.Month(), day.Day(), 0, 0, 0, 0, day.Location())
-}
-
-// spanDays returns the local calendar days an event touches. The end time
-// is treated as exclusive. An event that ends exactly at midnight then does
-// not count the following day. All-day events are stored as midnight-UTC
-// datestamps. Their span is derived from the UTC date, anchored at local
-// midnight to match the timed-event convention used by callers' window and
-// boundary comparisons. They then land on the correct day regardless of the
-// local timezone offset (mirrors eventCalendarDays).
-func spanDays(ev event.Event) []time.Time {
-	if ev.AllDay {
-		s := ev.StartTime.UTC()
-		e := ev.EndTime.UTC()
-		startDay := time.Date(s.Year(), s.Month(), s.Day(), 0, 0, 0, 0, time.Local)
-		var days []time.Time
-		for d := s; d.Before(e); d = d.AddDate(0, 0, 1) {
-			days = append(days, time.Date(d.Year(), d.Month(), d.Day(), 0, 0, 0, 0, time.Local))
-		}
-		if len(days) == 0 {
-			days = []time.Time{startDay}
-		}
-		return days
-	}
-	s := ev.StartTime.Local()
-	startDay := time.Date(s.Year(), s.Month(), s.Day(), 0, 0, 0, 0, s.Location())
-	e := ev.EndTime.Local()
-	if !e.After(s) {
-		return []time.Time{startDay}
-	}
-	var days []time.Time
-	for d := startDay; d.Before(e); d = d.AddDate(0, 0, 1) {
-		days = append(days, d)
-	}
-	if len(days) == 0 {
-		days = []time.Time{startDay}
-	}
-	return days
-}
-
-type FormatEventListOptions struct {
-	Events        []event.Event
-	CalendarNames map[int64]string
-	ShowHeader    bool
-	ShowAllDays   bool
-	From          time.Time
-	To            time.Time
-	// WeekdayWidth controls the weekday label width (1, 2, or 3 chars).
-	// Zero or out-of-range values default to 3.
-	WeekdayWidth int
-	// ShowWeekday controls whether the weekday label is displayed.
-	ShowWeekday bool
-	// ShowMonth controls whether the month label is displayed in the day prefix.
-	ShowMonth bool
-	// Verbose renders a richer time-rail view and suppresses empty days.
-	Verbose bool
-	// ShowID appends the event ID in text output.
-	ShowID bool
-	// ShowCalendar appends or prints the calendar name in text output.
-	ShowCalendar bool
-}
-
-type eventListDayEntry struct {
-	ev        event.Event
-	dayIndex  int
-	totalDays int
-}
-
-// formatWeekday returns a 1-, 2-, or 3-character English weekday label.
-func formatWeekday(t time.Time, width int) string {
-	full := t.Format("Mon")
-	switch width {
-	case 1, 2:
-		return full[:width]
-	default:
-		return full
-	}
-}
-
-func FormatEventList(opts FormatEventListOptions) string {
-	if len(opts.Events) == 0 && !opts.ShowAllDays {
-		return ""
-	}
-
-	weekdayWidth := opts.WeekdayWidth
-	if weekdayWidth < 1 || weekdayWidth > 3 {
-		weekdayWidth = 3
-	}
-
-	eventsByDay := make(map[string][]eventListDayEntry)
-	for _, ev := range opts.Events {
-		days := spanDays(ev)
-		total := len(days)
-		for i, d := range days {
-			key := d.Format("2006-01-02")
-			eventsByDay[key] = append(eventsByDay[key], eventListDayEntry{
-				ev:        ev,
-				dayIndex:  i + 1,
-				totalDays: total,
-			})
-		}
-	}
-
-	// Sort each day's entries so continuations of earlier events (which are
-	// active from the start of the day) appear before events starting later.
-	for k, entries := range eventsByDay {
-		day, _ := time.ParseInLocation("2006-01-02", k, time.Local)
-		sort.SliceStable(entries, func(a, b int) bool {
-			return effectiveStartOnDay(entries[a].ev, day, entries[a].dayIndex).
-				Before(effectiveStartOnDay(entries[b].ev, day, entries[b].dayIndex))
-		})
-		eventsByDay[k] = entries
-	}
-
-	months := make(map[string][]string)
-	var monthOrder []string
-
-	addDay := func(d time.Time) {
-		monthKey := d.Format("2006-01")
-		dayKey := d.Format("2006-01-02")
-		if _, exists := months[monthKey]; !exists {
-			monthOrder = append(monthOrder, monthKey)
-		}
-		months[monthKey] = append(months[monthKey], dayKey)
-	}
-
-	if opts.ShowAllDays && !opts.From.IsZero() && !opts.To.IsZero() {
-		from := time.Date(opts.From.Year(), opts.From.Month(), opts.From.Day(), 0, 0, 0, 0, time.Local)
-		to := time.Date(opts.To.Year(), opts.To.Month(), opts.To.Day(), 0, 0, 0, 0, time.Local)
-		for d := from; d.Before(to); d = d.AddDate(0, 0, 1) {
-			addDay(d)
-		}
-	} else {
-		seen := make(map[string]bool)
-		for _, ev := range opts.Events {
-			for _, d := range spanDays(ev) {
-				dayKey := d.Format("2006-01-02")
-				if !seen[dayKey] {
-					seen[dayKey] = true
-					addDay(d)
-				}
-			}
-		}
-	}
-
-	var out strings.Builder
-	firstVerboseDay := true
-	for _, monthKey := range monthOrder {
-		if opts.ShowHeader {
-			t, _ := time.Parse("2006-01", monthKey)
-			out.WriteString(lipgloss.NewStyle().Bold(true).Render(t.Format("January 2006")))
-			out.WriteString("\n\n")
-		}
-
-		for _, dayKey := range months[monthKey] {
-			dayEvents := eventsByDay[dayKey]
-			d, _ := time.Parse("2006-01-02", dayKey)
-			dayPrefix := d.Format("02")
-			if opts.ShowWeekday {
-				dayPrefix += " " + formatWeekday(d, weekdayWidth)
-			}
-			if opts.ShowMonth {
-				dayPrefix = d.Format("Jan") + " " + dayPrefix
-			}
-
-			if opts.Verbose {
-				if len(dayEvents) == 0 {
-					continue
-				}
-				if !firstVerboseDay {
-					out.WriteByte('\n')
-				}
-				out.WriteString(dayPrefix)
-				out.WriteByte('\n')
-				out.WriteString(strings.Repeat("-", len(dayPrefix)))
-				out.WriteByte('\n')
-				for _, entry := range dayEvents {
-					writeVerboseEventListEntry(&out, entry, opts)
-				}
-				firstVerboseDay = false
-				continue
-			}
-
-			if len(dayEvents) == 0 {
-				out.WriteString(dayPrefix)
-				out.WriteByte('\n')
-				continue
-			}
-
-			continuation := strings.Repeat(" ", len(dayPrefix))
-			for i, entry := range dayEvents {
-				if i == 0 {
-					out.WriteString(dayPrefix)
-				} else {
-					out.WriteString(continuation)
-				}
-				out.WriteString(" ")
-				out.WriteString(formatTimeColumnMulti(entry.ev, entry.dayIndex, entry.totalDays))
-				out.WriteString("  ")
-				out.WriteString(compactEventLabel(entry, opts))
-				out.WriteByte('\n')
-			}
-		}
-
-		if opts.ShowHeader {
-			out.WriteByte('\n')
-		}
-	}
-
-	return out.String()
-}
-
-func verboseTimeRailLabel(ev event.Event, dayIndex int) string {
-	switch {
-	case ev.AllDay:
-		return "all day"
-	case dayIndex == 1:
-		return ev.StartTime.Local().Format("15:04")
-	default:
-		return "00:00"
-	}
-}
-
-func verboseContinuationLabel(ev event.Event, dayIndex, totalDays int) string {
-	if totalDays <= 1 || ev.AllDay {
-		return ""
-	}
-	switch dayIndex {
-	case 1:
-		return "ends " + ev.EndTime.Local().Format("Mon, Jan 2 15:04")
-	case totalDays:
-		return "until " + ev.EndTime.Local().Format("15:04")
-	default:
-		return "continues"
-	}
-}
-
-func compactEventLabel(entry eventListDayEntry, opts FormatEventListOptions) string {
-	label := entry.ev.Title
-	if opts.ShowID && entry.ev.ID > 0 {
-		label += fmt.Sprintf(" (%d)", entry.ev.ID)
-	}
-	if entry.totalDays > 1 {
-		label += fmt.Sprintf(" (day %d/%d)", entry.dayIndex, entry.totalDays)
-	}
-	if opts.ShowCalendar {
-		if calendarLabel := formatCalendarLabel(entry.ev.CalendarID, opts.CalendarNames); calendarLabel != "" {
-			label += " [" + calendarLabel + "]"
-		}
-	}
-	return label
-}
-
-func writeVerboseEventListEntry(out *strings.Builder, entry eventListDayEntry, opts FormatEventListOptions) {
-	title := entry.ev.Title
-	if opts.ShowID && entry.ev.ID > 0 {
-		title = fmt.Sprintf("%s (%d)", title, entry.ev.ID)
-	}
-	if entry.totalDays > 1 {
-		title = fmt.Sprintf("%s (day %d/%d)", title, entry.dayIndex, entry.totalDays)
-	}
-
-	fmt.Fprintf(out, "%-7s | %s\n", verboseTimeRailLabel(entry.ev, entry.dayIndex), title)
-	if entry.ev.Location != "" {
-		fmt.Fprintf(out, "%7s | %s\n", "", entry.ev.Location)
-	}
-	if entry.ev.Description != "" {
-		fmt.Fprintf(out, "%7s | %s\n", "", entry.ev.Description)
-	}
-	if metadata := verboseMetadataLine(entry.ev, opts); metadata != "" {
-		fmt.Fprintf(out, "%7s | %s\n", "", metadata)
-	}
-	if continuation := verboseContinuationLabel(entry.ev, entry.dayIndex, entry.totalDays); continuation != "" {
-		fmt.Fprintf(out, "%7s | %s\n", "", continuation)
-	}
-}
-
-func formatCalendarLabel(calendarID int64, calendarNames map[int64]string) string {
-	if calendarID <= 0 {
-		return ""
-	}
-	if name := calendarNames[calendarID]; name != "" {
-		return name
-	}
-	return fmt.Sprintf("%d", calendarID)
-}
-
-func verboseMetadataLine(ev event.Event, opts FormatEventListOptions) string {
-	if !opts.ShowCalendar {
-		return ""
-	}
-	if calendarLabel := formatCalendarLabel(ev.CalendarID, opts.CalendarNames); calendarLabel != "" {
-		return "Calendar: " + calendarLabel
-	}
-	return ""
-}
 
 // CalendarEvent is the render-only view of an event inside the month grid.
 // Callers resolve colors and other domain data before they pass these in.
@@ -368,356 +27,6 @@ type CalendarEvent struct {
 	Day       time.Time // local day the event should render on
 	StartTime time.Time
 	EndTime   time.Time
-}
-
-type CalendarOptions struct {
-	Month            time.Time
-	Events           []CalendarEvent
-	Today            time.Time
-	Selected         time.Time
-	WeekStartsOn     time.Weekday
-	Width            int
-	Height           int
-	ShowHeader       bool
-	ShowAdjacentDays bool
-	ShowWeekNumbers  bool
-	// SelectedColor, when non-nil, redraws the selected cell's borders in
-	// this color. Use the theme's text color for a "highlighted cursor" look.
-	SelectedColor color.Color
-}
-
-// isoWeekForRow returns the ISO week number of the Thursday within the
-// week row that starts at anchor+week*7 (with weekStart as the first
-// column). Thursday is always in the same ISO week as the row's dates.
-func isoWeekForRow(anchor time.Time, weekStart time.Weekday, week int) int {
-	thursdayOffset := (int(time.Thursday) - int(weekStart) + 7) % 7
-	d := anchor.AddDate(0, 0, week*7+thursdayOffset)
-	_, w := d.ISOWeek()
-	return w
-}
-
-// weekdayOffset returns how many days d falls after weekStart, in 0..6.
-func weekdayOffset(d time.Time, weekStart time.Weekday) int {
-	return (int(d.Weekday()) - int(weekStart) + 7) % 7
-}
-
-// calendarGridAnchor returns the date of the top-left cell of the calendar
-// grid for the given month and week start.
-func calendarGridAnchor(month time.Time, weekStart time.Weekday) time.Time {
-	first := time.Date(month.Year(), month.Month(), 1, 0, 0, 0, 0, time.Local)
-	return first.AddDate(0, 0, -weekdayOffset(first, weekStart))
-}
-
-// distributeCells splits avail into n cell sizes each >= minSize. It spreads
-// the remainder across cells at the start so the row/column sums to avail.
-func distributeCells(avail, n, minSize int) []int {
-	base := max(avail/n, minSize)
-	rem := max(avail-base*n, 0)
-	sizes := make([]int, n)
-	for i := range n {
-		sizes[i] = base
-		if i < rem {
-			sizes[i]++
-		}
-	}
-	return sizes
-}
-
-// calendarGridSizes returns the per-column widths, per-row heights, and the
-// number of preamble lines above the 6-row table for a Width×Height month grid.
-func calendarGridSizes(width, height int, showHeader bool) (cellWs, cellHs []int, preambleLines int) {
-	// Preamble lines above the table: title + blank (optional) + weekday row.
-	preambleLines = 1
-	if showHeader {
-		preambleLines += 2
-	}
-	// Table overhead: 8 vertical borders between 7 columns.
-	cellWs = distributeCells(width-8, 7, 6)
-	// Table overhead: top + bottom + 5 inter-row borders = 7 chrome lines
-	// above the 6 week rows (no header row — weekdays render as preamble).
-	cellHs = distributeCells(height-preambleLines-7, 6, 2)
-	return cellWs, cellHs, preambleLines
-}
-
-// Calendar renders a full-size month grid that fills Width×Height.
-// Returns "" if Width or Height is zero.
-func Calendar(opts CalendarOptions) string {
-	if opts.Width <= 0 || opts.Height <= 0 {
-		return ""
-	}
-
-	anchor := calendarGridAnchor(opts.Month, opts.WeekStartsOn)
-	first := time.Date(opts.Month.Year(), opts.Month.Month(), 1, 0, 0, 0, 0, time.Local)
-
-	eventsByDay := make(map[string][]CalendarEvent)
-	for _, ev := range opts.Events {
-		key := ev.Day.Format("2006-01-02")
-		eventsByDay[key] = append(eventsByDay[key], ev)
-	}
-
-	todayKey := ""
-	if !opts.Today.IsZero() {
-		todayKey = opts.Today.Local().Format("2006-01-02")
-	}
-	selectedKey := ""
-	if !opts.Selected.IsZero() {
-		selectedKey = opts.Selected.Local().Format("2006-01-02")
-	}
-
-	cellWs, cellHs, _ := calendarGridSizes(opts.Width, opts.Height, opts.ShowHeader)
-
-	rows := make([][]string, 6)
-	for week := range 6 {
-		row := make([]string, 7)
-		weekLabel := ""
-		if opts.ShowWeekNumbers {
-			weekLabel = fmt.Sprintf("W%d", isoWeekForRow(anchor, opts.WeekStartsOn, week))
-		}
-		for col := range 7 {
-			d := anchor.AddDate(0, 0, week*7+col)
-			dayKey := d.Format("2006-01-02")
-			inMonth := d.Month() == first.Month() && d.Year() == first.Year()
-
-			label := ""
-			if col == 0 {
-				label = weekLabel
-			}
-
-			if !inMonth && !opts.ShowAdjacentDays {
-				row[col] = blankCellWithWeekLabel(cellWs[col], cellHs[week], label)
-				continue
-			}
-			row[col] = buildCalendarCell(d, dayKey == todayKey, dayKey == selectedKey, inMonth, eventsByDay[dayKey], cellWs[col], cellHs[week], label)
-		}
-		rows[week] = row
-	}
-
-	t := table.New().
-		Rows(rows...).
-		Border(lipgloss.RoundedBorder()).
-		BorderRow(true).
-		BorderStyle(lipgloss.NewStyle().Faint(true)).
-		StyleFunc(func(_, col int) lipgloss.Style {
-			return lipgloss.NewStyle().Width(cellWs[col]).Padding(0, 0)
-		})
-
-	rendered := t.Render()
-	if opts.SelectedColor != nil && !opts.Selected.IsZero() {
-		sr, sc := findCellIndex(anchor, opts.Selected)
-		if sr >= 0 {
-			rendered = highlightCellBorder(rendered, sr, sc, cellWs, cellHs, opts.SelectedColor)
-		}
-	}
-
-	var out strings.Builder
-	if opts.ShowHeader {
-		out.WriteString(lipgloss.NewStyle().Bold(true).Width(opts.Width).Align(lipgloss.Left).Render(first.Format("January 2006")))
-		out.WriteString("\n\n")
-	}
-	out.WriteString(renderWeekdayRow(anchor, cellWs))
-	out.WriteString("\n")
-	out.WriteString(rendered)
-	return out.String()
-}
-
-// renderWeekdayRow returns a single-line row of centered, faint weekday
-// labels whose columns align with the calendar table below. The row pads
-// with a space at the start, end, and inner where the table's vertical
-// borders would sit so widths match exactly.
-func renderWeekdayRow(anchor time.Time, cellWs []int) string {
-	var b strings.Builder
-	b.WriteString(" ")
-	for i := range 7 {
-		if i > 0 {
-			b.WriteString(" ")
-		}
-		label := strings.ToLower(anchor.AddDate(0, 0, i).Format("Mon"))
-		b.WriteString(lipgloss.NewStyle().
-			Width(cellWs[i]).
-			Align(lipgloss.Center).
-			Faint(true).
-			Render(label))
-	}
-	b.WriteString(" ")
-	return b.String()
-}
-
-func findCellIndex(anchor, d time.Time) (int, int) {
-	target := d.Local().Format("2006-01-02")
-	for week := range 6 {
-		for col := range 7 {
-			cell := anchor.AddDate(0, 0, week*7+col)
-			if cell.Format("2006-01-02") == target {
-				return week, col
-			}
-		}
-	}
-	return -1, -1
-}
-
-func highlightCellBorder(rendered string, sr, sc int, cellWs, cellHs []int, c color.Color) string {
-	leftC := sc
-	for i := range sc {
-		leftC += cellWs[i]
-	}
-	rightC := leftC + cellWs[sc] + 1
-
-	topL := 0
-	for i := range sr {
-		topL += cellHs[i] + 1
-	}
-	botL := topL + cellHs[sr] + 1
-
-	style := lipgloss.NewStyle().Foreground(c).Bold(true).Faint(false)
-
-	lines := strings.Split(rendered, "\n")
-	for y := topL; y <= botL && y >= 0 && y < len(lines); y++ {
-		if y == topL {
-			lines[y] = substituteAtVisPos(lines[y], map[int]rune{leftC: '╭', rightC: '╮'})
-			lines[y] = lipgloss.StyleRanges(lines[y],
-				lipgloss.NewRange(leftC, rightC+1, style))
-			continue
-		}
-		if y == botL {
-			lines[y] = substituteAtVisPos(lines[y], map[int]rune{leftC: '╰', rightC: '╯'})
-			lines[y] = lipgloss.StyleRanges(lines[y],
-				lipgloss.NewRange(leftC, rightC+1, style))
-			continue
-		}
-		lines[y] = lipgloss.StyleRanges(lines[y],
-			lipgloss.NewRange(leftC, leftC+1, style),
-			lipgloss.NewRange(rightC, rightC+1, style))
-	}
-	return strings.Join(lines, "\n")
-}
-
-func substituteAtVisPos(line string, subs map[int]rune) string {
-	if len(subs) == 0 {
-		return line
-	}
-	var out strings.Builder
-	out.Grow(len(line))
-	vis := 0
-	for i := 0; i < len(line); {
-		if line[i] == 0x1b && i+1 < len(line) && line[i+1] == '[' {
-			j := i + 2
-			for j < len(line) {
-				b := line[j]
-				if b >= 0x40 && b <= 0x7e {
-					j++
-					break
-				}
-				j++
-			}
-			out.WriteString(line[i:j])
-			i = j
-			continue
-		}
-		r, size := utf8.DecodeRuneInString(line[i:])
-		if repl, ok := subs[vis]; ok {
-			out.WriteRune(repl)
-		} else {
-			out.WriteRune(r)
-		}
-		vis++
-		i += size
-	}
-	return out.String()
-}
-
-func blankCellWithWeekLabel(w, h int, weekLabel string) string {
-	line := strings.Repeat(" ", w)
-	lines := make([]string, h)
-	for i := range h {
-		lines[i] = line
-	}
-	if weekLabel != "" && h > 0 {
-		lines[0] = renderWeekLabelLine(weekLabel, "", w)
-	}
-	return strings.Join(lines, "\n")
-}
-
-// renderWeekLabelLine composes the first line of a calendar cell: the
-// faint "Wnn" label on the left, and the day-number block (already styled)
-// right-aligned. It is padded so the whole line is exactly cellW wide.
-func renderWeekLabelLine(weekLabel, dayRendered string, cellW int) string {
-	faintLabel := lipgloss.NewStyle().Faint(true).Render(weekLabel)
-	labelW := lipgloss.Width(faintLabel)
-	dayW := lipgloss.Width(dayRendered)
-	pad := max(cellW-labelW-dayW, 0)
-	return faintLabel + strings.Repeat(" ", pad) + dayRendered
-}
-
-func buildCalendarCell(d time.Time, isToday, isSelected, inMonth bool, events []CalendarEvent, cellW, cellH int, weekLabel string) string {
-	dayNum := fmt.Sprintf("%d", d.Day())
-
-	numStyle := lipgloss.NewStyle()
-	switch {
-	case isToday:
-		// Explicit fg/bg rather than Reverse(true): some terminal themes
-		// (e.g. Omarchy light themes with OSC-11 cream backgrounds) don't
-		// swap reliably between the ANSI 0/7 palette and the OSC-10/11
-		// defaults, which can collapse Reverse to fg==bg.
-		numStyle = numStyle.
-			Background(activeTheme.Today).
-			Foreground(activeTheme.Surface).
-			Bold(true).
-			Padding(0, 1)
-	case isSelected:
-		// Paint the selected day's number in the same accent used by the
-		// cell-border highlight, so the cursor day pops even when the
-		// border alone is subtle against the terminal bg.
-		numStyle = numStyle.Foreground(activeTheme.Primary).Bold(true)
-	case !inMonth:
-		numStyle = numStyle.Faint(true)
-	}
-
-	rendered := numStyle.Render(dayNum)
-	var numLine string
-	if weekLabel != "" {
-		numLine = renderWeekLabelLine(weekLabel, rendered, cellW)
-	} else {
-		padW := max(cellW-lipgloss.Width(rendered), 0)
-		numLine = strings.Repeat(" ", padW) + rendered
-	}
-
-	maxEventLines := cellH - 1
-	pills := make([]string, 0, maxEventLines)
-	overflow := 0
-	for i, ev := range events {
-		if i >= maxEventLines {
-			overflow = len(events) - maxEventLines + 1
-			break
-		}
-		pills = append(pills, renderEventPill(ev, cellW, !inMonth))
-	}
-	if overflow > 0 && len(pills) > 0 {
-		pills[len(pills)-1] = lipgloss.NewStyle().Faint(true).
-			Width(cellW).Render(fmt.Sprintf(" +%d more", overflow))
-	}
-
-	lines := make([]string, 0, cellH)
-	lines = append(lines, numLine)
-	lines = append(lines, pills...)
-	blank := strings.Repeat(" ", cellW)
-	for len(lines) < cellH {
-		lines = append(lines, blank)
-	}
-	return strings.Join(lines, "\n")
-}
-
-type WeekOptions struct {
-	WeekStart       time.Time
-	Events          []CalendarEvent
-	Today           time.Time
-	Selected        time.Time
-	Width           int
-	Height          int
-	ShowHeader      bool
-	ShowWeekNumbers bool
-	SelectedColor   color.Color
-	ScrollOffset    int
-	LinesPerHour    int
 }
 
 type placedEvent struct {
@@ -783,21 +92,6 @@ func resolveOverlaps(placed []placedEvent) {
 	}
 }
 
-func calcWeekColWidths(width int) []int {
-	separators := 8
-	availW := max(width-timeLabelWidth-separators, 7)
-	colW := availW / 7
-	remW := availW - colW*7
-	colWs := make([]int, 7)
-	for i := range 7 {
-		colWs[i] = colW
-		if i < remW {
-			colWs[i]++
-		}
-	}
-	return colWs
-}
-
 func placeEvents(events []CalendarEvent, colFn func(CalendarEvent) int, lph int) []placedEvent {
 	var placed []placedEvent
 	totalRows := totalHours * lph
@@ -828,297 +122,6 @@ func placeEvents(events []CalendarEvent, colFn func(CalendarEvent) int, lph int)
 		})
 	}
 	return placed
-}
-
-func placeWeekEvents(events []CalendarEvent, anchor time.Time, lph int) []placedEvent {
-	return placeEvents(events, func(ev CalendarEvent) int {
-		return findWeekCol(anchor, ev.Day)
-	}, lph)
-}
-
-func WeekGrid(opts WeekOptions) string {
-	if opts.Width <= 0 || opts.Height <= 0 {
-		return ""
-	}
-
-	lph := opts.LinesPerHour
-	if lph < 1 {
-		lph = defaultLinesPerHour
-	}
-	totalRows := totalHours * lph
-	anchor := opts.WeekStart
-
-	todayKey := ""
-	if !opts.Today.IsZero() {
-		todayKey = opts.Today.Local().Format("2006-01-02")
-	}
-	selectedKey := ""
-	selectedCol := -1
-	if !opts.Selected.IsZero() {
-		selectedKey = opts.Selected.Local().Format("2006-01-02")
-		selectedCol = findWeekCol(anchor, opts.Selected)
-	}
-
-	colWs := calcWeekColWidths(opts.Width)
-	placed := placeWeekEvents(opts.Events, anchor, lph)
-	resolveOverlaps(placed)
-
-	allDayRows := max(weekAllDayRowCount(opts.Events, anchor), 1)
-
-	headerLines := 0
-	if opts.ShowHeader {
-		headerLines = 2
-	}
-	fixedLines := headerLines + 1 + 1 + allDayRows + 1
-	viewportHeight := max(opts.Height-fixedLines, 1)
-
-	maxScroll := max(totalRows-viewportHeight, 0)
-	scrollOffset := min(max(opts.ScrollOffset, 0), maxScroll)
-
-	// Reserve one line for the bottom rule when end of day is visible.
-	showBottomRule := scrollOffset+viewportHeight >= totalRows
-	if showBottomRule && viewportHeight > 1 {
-		viewportHeight--
-	}
-
-	now := time.Now().Local()
-	nowRow := now.Hour()*lph + now.Minute()*lph/60
-	nowTimeLabel := now.Format("15:04")
-	nowCol := findWeekCol(anchor, now)
-	nowHasLine := nowCol >= 0
-
-	faint := lipgloss.NewStyle().Faint(true)
-	faintSep := faint.Render("│")
-	nowStyle := lipgloss.NewStyle().Foreground(ActiveTheme().Today).Bold(true)
-	nowSep := nowStyle.Render("│")
-
-	var selSep string
-	if opts.SelectedColor != nil && selectedCol >= 0 {
-		selStyle := lipgloss.NewStyle().Foreground(opts.SelectedColor).Bold(true).Faint(false)
-		selSep = selStyle.Render("│")
-	}
-
-	var out strings.Builder
-
-	if opts.ShowHeader {
-		endDay := anchor.AddDate(0, 0, 6)
-		var title string
-		if anchor.Month() == endDay.Month() {
-			title = fmt.Sprintf("%s %d – %d, %d",
-				anchor.Format("January"), anchor.Day(), endDay.Day(), anchor.Year())
-		} else if anchor.Year() == endDay.Year() {
-			title = fmt.Sprintf("%s %d – %s %d, %d",
-				anchor.Format("Jan"), anchor.Day(), endDay.Format("Jan"), endDay.Day(), anchor.Year())
-		} else {
-			title = fmt.Sprintf("%s %d, %d – %s %d, %d",
-				anchor.Format("Jan"), anchor.Day(), anchor.Year(), endDay.Format("Jan"), endDay.Day(), endDay.Year())
-		}
-		out.WriteString(lipgloss.NewStyle().Bold(true).Width(opts.Width).Align(lipgloss.Left).Render(title))
-		out.WriteString("\n\n")
-	}
-
-	weekNumLabel := ""
-	if opts.ShowWeekNumbers {
-		thursdayOffset := (int(time.Thursday) - int(anchor.Weekday()) + 7) % 7
-		_, wn := anchor.AddDate(0, 0, thursdayOffset).ISOWeek()
-		weekNumLabel = fmt.Sprintf("W%d", wn)
-	}
-	out.WriteString(renderWeekColumnHeaders(anchor, colWs, todayKey, selectedKey, opts.SelectedColor, weekNumLabel))
-	out.WriteString("\n")
-
-	out.WriteString(renderWeekHRule(colWs, "┌", "┬", "", true, selectedCol, opts.SelectedColor))
-	out.WriteString("\n")
-
-	out.WriteString(renderWeekAllDayRows(opts.Events, anchor, colWs, allDayRows, selectedCol, opts.SelectedColor))
-
-	out.WriteString(renderWeekHRule(colWs, "├", "┼", "╮", true, selectedCol, opts.SelectedColor))
-	out.WriteString("\n")
-
-	for row := scrollOffset; row < scrollOffset+viewportHeight && row < totalRows; row++ {
-		if row > scrollOffset {
-			out.WriteString("\n")
-		}
-
-		out.WriteString(renderTimeLabel(row, lph, row == nowRow, nowTimeLabel))
-
-		for i := 0; i <= 7; i++ {
-			nowBorder := nowHasLine && row == nowRow && i == nowCol
-			highlighted := selSep != "" && (i == selectedCol || i == selectedCol+1)
-			if nowBorder {
-				out.WriteString(nowSep)
-			} else if highlighted {
-				out.WriteString(selSep)
-			} else {
-				out.WriteString(faintSep)
-			}
-
-			if i < 7 {
-				matches := findPlacedEvents(placed, row, i)
-				if len(matches) > 0 {
-					out.WriteString(renderOverlappingCells(matches, row, colWs[i]))
-				} else if nowHasLine && row == nowRow && i == nowCol {
-					out.WriteString(nowStyle.Render(strings.Repeat("─", colWs[i])))
-				} else {
-					out.WriteString(strings.Repeat(" ", colWs[i]))
-				}
-			}
-		}
-	}
-
-	if showBottomRule {
-		out.WriteString("\n")
-		out.WriteString(renderWeekHRule(colWs, "╰", "┴", "╯", false, selectedCol, opts.SelectedColor))
-	}
-
-	return out.String()
-}
-
-func renderWeekColumnHeaders(anchor time.Time, colWs []int, todayKey, selectedKey string, selectedColor color.Color, weekNumLabel string) string {
-	var b strings.Builder
-	if weekNumLabel != "" {
-		b.WriteString(lipgloss.NewStyle().Faint(true).Width(timeLabelWidth).Align(lipgloss.Center).Render(weekNumLabel))
-	} else {
-		b.WriteString(strings.Repeat(" ", timeLabelWidth))
-	}
-	b.WriteString(" ")
-	for i := range 7 {
-		if i > 0 {
-			b.WriteString(" ")
-		}
-		d := anchor.AddDate(0, 0, i)
-		dayKey := d.Format("2006-01-02")
-		dayName := strings.ToLower(d.Format("Mon"))
-		dayNum := fmt.Sprintf("%d", d.Day())
-		style := lipgloss.NewStyle().Faint(true)
-		numStyle := lipgloss.NewStyle().Faint(true)
-		if dayKey == todayKey {
-			// Today number is always the filled pill — the cursor never
-			// overrides today's identity on the number itself.
-			style = style.Faint(false).Bold(true)
-			numStyle = numStyle.
-				Faint(false).
-				Bold(true).
-				Background(activeTheme.Today).
-				Foreground(activeTheme.Surface).
-				Padding(0, 1)
-		}
-		if dayKey == selectedKey && selectedColor != nil {
-			// The weekday label picks up the cursor accent in both cases,
-			// but we skip the number when it's also today so the filled
-			// pill keeps its today colors.
-			style = style.Foreground(selectedColor).Bold(true).Faint(false)
-			if dayKey != todayKey {
-				numStyle = numStyle.Foreground(selectedColor).Bold(true).Faint(false)
-			}
-		}
-		label := style.Render(dayName) + " " + numStyle.Render(dayNum)
-		colStyle := lipgloss.NewStyle().Width(colWs[i]).Align(lipgloss.Center)
-		b.WriteString(colStyle.Render(label))
-	}
-	b.WriteString(" ")
-	return b.String()
-}
-
-func renderWeekHRule(colWs []int, left, mid, right string, timeCol bool, selectedCol int, selectedColor color.Color) string {
-	faint := lipgloss.NewStyle().Faint(true)
-	var selStyle lipgloss.Style
-	hasSel := selectedColor != nil && selectedCol >= 0
-	if hasSel {
-		selStyle = lipgloss.NewStyle().Foreground(selectedColor).Bold(true).Faint(false)
-	}
-
-	renderJunction := func(s string, sepIdx int) string {
-		if hasSel && (sepIdx == selectedCol || sepIdx == selectedCol+1) {
-			return selStyle.Render(s)
-		}
-		return faint.Render(s)
-	}
-
-	var b strings.Builder
-	if timeCol {
-		b.WriteString(faint.Render("────────"))
-		b.WriteString(renderJunction(left, 0))
-	} else {
-		b.WriteString("        ")
-		b.WriteString(renderJunction(left, 0))
-	}
-	for i, w := range colWs {
-		seg := strings.Repeat("─", w)
-		if hasSel && i == selectedCol {
-			b.WriteString(selStyle.Render(seg))
-		} else {
-			b.WriteString(faint.Render(seg))
-		}
-		if i < len(colWs)-1 {
-			b.WriteString(renderJunction(mid, i+1))
-		}
-	}
-	b.WriteString(renderJunction(right, len(colWs)))
-	return b.String()
-}
-
-func weekAllDayRowCount(events []CalendarEvent, anchor time.Time) int {
-	maxPerCol := 0
-	for col := range 7 {
-		d := anchor.AddDate(0, 0, col)
-		dayKey := d.Format("2006-01-02")
-		count := 0
-		for _, ev := range events {
-			if ev.AllDay && ev.Day.Format("2006-01-02") == dayKey {
-				count++
-			}
-		}
-		if count > maxPerCol {
-			maxPerCol = count
-		}
-	}
-	return maxPerCol
-}
-
-func renderWeekAllDayRows(events []CalendarEvent, anchor time.Time, colWs []int, numRows int, selectedCol int, selectedColor color.Color) string {
-	eventsByCol := make([][]CalendarEvent, 7)
-	for _, ev := range events {
-		if !ev.AllDay {
-			continue
-		}
-		col := findWeekCol(anchor, ev.Day)
-		if col >= 0 {
-			eventsByCol[col] = append(eventsByCol[col], ev)
-		}
-	}
-
-	faint := lipgloss.NewStyle().Faint(true)
-	faintSep := faint.Render("│")
-	var selSep string
-	if selectedColor != nil && selectedCol >= 0 {
-		selStyle := lipgloss.NewStyle().Foreground(selectedColor).Bold(true).Faint(false)
-		selSep = selStyle.Render("│")
-	}
-
-	var out strings.Builder
-	for row := range numRows {
-		if row == 0 {
-			out.WriteString(faint.Render("All day") + " ")
-		} else {
-			out.WriteString("        ")
-		}
-		for i := range 7 {
-			highlighted := selSep != "" && (i == selectedCol || i == selectedCol+1)
-			if highlighted {
-				out.WriteString(selSep)
-			} else {
-				out.WriteString(faintSep)
-			}
-			if row < len(eventsByCol[i]) {
-				out.WriteString(renderEventPill(eventsByCol[i][row], colWs[i], false))
-			} else {
-				out.WriteString(strings.Repeat(" ", colWs[i]))
-			}
-		}
-		out.WriteString("\n")
-	}
-
-	return out.String()
 }
 
 func findPlacedEvents(placed []placedEvent, row, col int) []placedEvent {
@@ -1213,18 +216,6 @@ func renderOverlappingCells(matches []placedEvent, row, totalWidth int) string {
 	return b.String()
 }
 
-func findWeekCol(anchor, d time.Time) int {
-	// d is already in the correct display timezone (local for timed events,
-	// UTC for all-day events via eventDay), so don't call .Local() again.
-	target := d.Format("2006-01-02")
-	for col := range 7 {
-		if anchor.AddDate(0, 0, col).Format("2006-01-02") == target {
-			return col
-		}
-	}
-	return -1
-}
-
 func renderTimeLabel(row, lph int, isNowRow bool, nowTimeLabel string) string {
 	if isNowRow {
 		s := lipgloss.NewStyle().Foreground(ActiveTheme().Today).Bold(true)
@@ -1250,180 +241,4 @@ func renderEventPill(ev CalendarEvent, cellW int, faint bool) string {
 	fg := oklch.ContrastingFg(bg)
 	return lipgloss.NewStyle().Background(bg).Foreground(fg).
 		Width(cellW).Render(text)
-}
-
-type DayOptions struct {
-	Day          time.Time
-	Events       []CalendarEvent
-	Today        time.Time
-	Width        int
-	Height       int
-	ShowHeader   bool
-	ScrollOffset int
-	LinesPerHour int
-}
-
-func DayGrid(opts DayOptions) string {
-	if opts.Width <= 0 || opts.Height <= 0 {
-		return ""
-	}
-
-	lph := opts.LinesPerHour
-	if lph < 1 {
-		lph = defaultLinesPerHour
-	}
-	totalRows := totalHours * lph
-	day := opts.Day
-
-	colWidth := max(opts.Width-timeLabelWidth-2, 1)
-
-	placed := placeDayEvents(opts.Events, day, lph)
-	resolveOverlaps(placed)
-
-	allDayRows := max(dayAllDayCount(opts.Events, day), 1)
-
-	headerLines := 0
-	if opts.ShowHeader {
-		headerLines = 2
-	}
-	fixedLines := headerLines + 1 + allDayRows + 1
-	viewportHeight := max(opts.Height-fixedLines, 1)
-
-	maxScroll := max(totalRows-viewportHeight, 0)
-	scrollOffset := min(max(opts.ScrollOffset, 0), maxScroll)
-
-	showBottomRule := scrollOffset+viewportHeight >= totalRows
-	if showBottomRule && viewportHeight > 1 {
-		viewportHeight--
-	}
-
-	now := time.Now().Local()
-	nowRow := now.Hour()*lph + now.Minute()*lph/60
-	nowTimeLabel := now.Format("15:04")
-	isToday := day.Format("2006-01-02") == now.Format("2006-01-02")
-
-	faint := lipgloss.NewStyle().Faint(true)
-	faintSep := faint.Render("│")
-	nowStyle := lipgloss.NewStyle().Foreground(ActiveTheme().Today).Bold(true)
-	nowSep := nowStyle.Render("│")
-
-	var out strings.Builder
-
-	if opts.ShowHeader {
-		title := day.Format("Monday, January 2, 2006")
-		out.WriteString(lipgloss.NewStyle().Bold(true).Width(opts.Width).Align(lipgloss.Left).Render(title))
-		out.WriteString("\n\n")
-	}
-
-	out.WriteString(renderDayHRule(colWidth, "┌", "", true))
-	out.WriteString("\n")
-
-	out.WriteString(renderDayAllDayRows(opts.Events, day, colWidth, allDayRows))
-
-	out.WriteString(renderDayHRule(colWidth, "├", "╮", true))
-	out.WriteString("\n")
-
-	for row := scrollOffset; row < scrollOffset+viewportHeight && row < totalRows; row++ {
-		if row > scrollOffset {
-			out.WriteString("\n")
-		}
-
-		isNowRow := isToday && row == nowRow
-		out.WriteString(renderTimeLabel(row, lph, isNowRow, nowTimeLabel))
-
-		if isNowRow {
-			out.WriteString(nowSep)
-		} else {
-			out.WriteString(faintSep)
-		}
-
-		matches := findPlacedEvents(placed, row, 0)
-		if len(matches) > 0 {
-			out.WriteString(renderOverlappingCells(matches, row, colWidth))
-		} else if isNowRow {
-			out.WriteString(nowStyle.Render(strings.Repeat("─", colWidth)))
-		} else {
-			out.WriteString(strings.Repeat(" ", colWidth))
-		}
-
-		if isNowRow {
-			out.WriteString(nowSep)
-		} else {
-			out.WriteString(faintSep)
-		}
-	}
-
-	if showBottomRule {
-		out.WriteString("\n")
-		out.WriteString(renderDayHRule(colWidth, "╰", "╯", false))
-	}
-
-	return out.String()
-}
-
-func placeDayEvents(events []CalendarEvent, day time.Time, lph int) []placedEvent {
-	dayKey := day.Format("2006-01-02")
-	return placeEvents(events, func(ev CalendarEvent) int {
-		if ev.Day.Format("2006-01-02") == dayKey {
-			return 0
-		}
-		return -1
-	}, lph)
-}
-
-func dayAllDayCount(events []CalendarEvent, day time.Time) int {
-	dayKey := day.Format("2006-01-02")
-	count := 0
-	for _, ev := range events {
-		if ev.AllDay && ev.Day.Format("2006-01-02") == dayKey {
-			count++
-		}
-	}
-	return count
-}
-
-func renderDayAllDayRows(events []CalendarEvent, day time.Time, colWidth int, numRows int) string {
-	dayKey := day.Format("2006-01-02")
-	var allDayEvents []CalendarEvent
-	for _, ev := range events {
-		if ev.AllDay && ev.Day.Format("2006-01-02") == dayKey {
-			allDayEvents = append(allDayEvents, ev)
-		}
-	}
-
-	faint := lipgloss.NewStyle().Faint(true)
-	faintSep := faint.Render("│")
-
-	var out strings.Builder
-	for row := range numRows {
-		if row == 0 {
-			out.WriteString(faint.Render("All day") + " ")
-		} else {
-			out.WriteString("        ")
-		}
-		out.WriteString(faintSep)
-		if row < len(allDayEvents) {
-			out.WriteString(renderEventPill(allDayEvents[row], colWidth, false))
-		} else {
-			out.WriteString(strings.Repeat(" ", colWidth))
-		}
-		out.WriteString("\n")
-	}
-	return out.String()
-}
-
-func renderDayHRule(colWidth int, left, right string, timeCol bool) string {
-	faint := lipgloss.NewStyle().Faint(true)
-	var b strings.Builder
-	if timeCol {
-		b.WriteString(faint.Render("────────"))
-	} else {
-		b.WriteString("        ")
-	}
-	b.WriteString(faint.Render(left))
-	b.WriteString(faint.Render(strings.Repeat("─", colWidth)))
-	if right != "" {
-		b.WriteString(faint.Render(right))
-	}
-	return b.String()
 }
